ingestion/pyproject.toml
[project]
name = "maa-ingest"
version = "0.1.0"
description = "S3 PDF ingestion worker for Maa books → Postgres + pgvector"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35",
    "openai>=1.54",
    "pgvector>=0.3",
    "Pillow>=10.0",
    "psycopg[binary]>=3.2",
    "pymupdf>=1.24",
    "pytesseract>=0.3.10",
    "python-dotenv>=1.0",
    "tiktoken>=0.8",
]

[project.scripts]
maa-ingest = "maa_ingest.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/maa_ingest"]

[tool.hatch.build.targets.sdist]
include = ["src/maa_ingest"]
